Move GaiaException to util.py to avoid circular imports
