+++ gaia/io/gaia_reader.py
@@ -5,8 +5,8 @@ from builtins import (
 from future.utils import with_metaclass
 
 from gaia.gaia_data import GaiaDataObject
-from gaia import GaiaException
 from gaia.util import (
+    GaiaException,
     MissingParameterError,
     MissingDataException,
     UnsupportedFormatException,
